[FIX] ERD 설계와 다르거나 누락된 옵션 수정

+++ src/main/java/org/websoso/WSSServer/domain/Keyword.java
@@ -21,7 +21,7 @@ public class Keyword {
     @Id
     @GeneratedValue(strategy = IDENTITY)
     @Column(nullable = false)
-    private Long keywordId;
+    private Integer keywordId;
 
     @Column(nullable = false)
     private KeywordCategory categoryName;
